Fix a problem with a dereffed target during clicks
The comparison is now checking whether a ref was set for target
which means that the event.target should actually be a value and
it should be the same as the referenced value.
If the reffed object is deleted and the new pointer_down event
has target set to None, then the previous comparison would not
be correct.