Event system (#258)
Event system for pygfx, with hierarchical event handling system, pointer capture and more.

The renderer acts as the virtual event root handler. Internally it will capture wgpu-py events and convert those into Event objects. Using the picking info, an optional target will be set whenever there is a WorldObject under the pointer. The event will be dispatched onto that target, up until the root, as long as the event bubbles and is not cancelled.

Please take a look at [events.py](examples/events.py) to see the event system in action.

Co-authored-by: Korijn van Golen <k.vangolen@clinicalgraphics.com>
Co-authored-by: Almar Klein <almar@almarklein.org>