Use wgsl saturate instrinsic when we can.
I searched for

```
rg "clamp\(.*0\.0, 1\.1 "
```

and it found a few occurances.

Honestly, this might be a micro optimization since the WGSL
documentation states that "saturate expands to clamp(x, 0.0, 1.0)"
but again, some hardware documentation lists that that the call to
"saturate" is free.

I know this is a style thing, so feel free to close.

I did this cleanup in my own codebase so I figured I would oblige.