Allow picking of background objects
So maybe I'm hacking too much and feel free to reject this patch, but
I'm overlaying one scene onto the main scene to create some kind of
picture in picture preview.

It would be really useful to interact with the background of the scene
since it indicates that the user has clicked on the picture in picture
(but maybe not on an object).

I haven't really gone through the codepath to make picking background
textures meaningful but this is already more than enough for me.

Otherwise, the pick just goes to the "object behind my picture in
picture".

Thanks for considering