Recall the last used font properties with set_text
I believe a common usecase is to generate the text geometry "once",

then keep updating it with a call to `set_text` without always thinking
about the font and other properties.

It would be good to start to "recall" the last used font properties,
style and weight. We currently drag along the font style and weight
every time we try to set the text in our application. We have code that
changes the displayed text often, similar to the stats counter, but we
would like it to maintain the same "text"

https://github.com/pygfx/pygfx/blob/2927712faeb867412a233b3efc8743ba1cc4cb11/pygfx/helpers/_stats.py#L123

Can become more commonplace.

Thanks for considering.