Provide options to better control font look
The general issues arise when redering text are the following:

- The glyph ref size of 48 is a little too small for some narrow fonts.
     - Proposal, we can provide a parameter to override it at the
       Geometry Level.
- The outline width always "adds" thickness to the font, given the low
  resolution of the above, this doesn't walways look good.
     - Proposal, provide an "inner-outline-width" parameter.

I demonstrate this with a font that is slightly narrower than Noto Sans
(Honestly, it took us a very long time to verify that the font was
different since they are so similar).

Let me know what you think.

Best,