Add image comparator example (#1019)
* Add image comparator example

Is there a better place for this kind of tool? I feel like it uses
text replace on the WGSL code and shouldn't really be encouraged by
users that don't want their code to break every other version.

Let me know if you think this tool would be a useful addition, I've
added it to my personal arsenal.

* refactor

* Add a big warning