Support marker rotation
I want to create a few arrows and whatnot, eventually I think this kind
of rotation will be useful for us as a whole.

The SDF is also the only way we have to "draw a circle perfectly".
However, I realize in all my enginuity that circles don't rotate!

- [ ] What should "auto" mean?
- [ ] I'm not sure what "vertex_map", is it meaningful for rotations?
- [ ] Which direction should positive rotate in?

It isn't quite what @MrLixm wanted, but should be able to do what they
wanted.

Closes https://github.com/pygfx/pygfx/issues/996